// LDC.Atlas.WebApp/Client/app/shared/enums/itparameter-documenttype.enum.ts
export enum ItParameterDocumentTypes {
    Purchase = 0,
    Sale = 1,
    Cost = 2,
    OtherSettlement = 3,
    Reversal = 4,
    GoodsAndCost = 5,
    PurchaseInvoice = 1,
    SalesInvoice = 2,
    CreditNote = 3,
    DebitNote = 4,
    CashPay = 5,
    CashReceipt = 6,
    TemporaryAdjustment = 7,
    MatchingCash = 8,
    RegularJournal = 9,
    FxDealJournal = 10,
    YearEndDocument = 11,
}